fix(examples): resolve all clippy warnings for CI inclusion
- Fix unused variable warnings by prefixing with underscore
- Fix unused import warnings by removing imports
- Fix dead_code warnings by adding #[allow(dead_code)] to structs
- Fix type_complexity warnings by adding #[allow(clippy::type_complexity)]
- Fix needless_range_loop warnings by adding #[allow] or refactoring
- Fix useless_vec by converting to arrays where appropriate
- Fix manual_strip by using strip_prefix
- Fix manual_clamp by using clamp()
- Fix if_same_then_else by simplifying conditionals
- Fix vec_init_then_push by using vec![] macro
- Update CI workflow to include examples in test and clippy jobs

All examples now build clean with -D warnings enabled.

+++ crates/haagenti-zstd/examples/decode_both.rs
@@ -107,7 +107,7 @@ fn decode_bitstream(bitstream: &[u8], seq_count: usize) {
         };
 
         let ll_value = get_ll_baseline(ll_code) + ll_extra as u32;
-        let of_value = (1u32 << of_code) + of_extra as u32;
+        let of_value = (1u32 << of_code) + of_extra;
         let ml_baseline = get_ml_baseline(ml_code);
         let ml_value = ml_baseline + ml_extra as u32;
